chore: extract common-type and c++23-backport specializations from StatePrinter.hpp
