+++ include/mimic++/Printer.hpp
@@ -14,5 +14,7 @@
 #include "mimic++/printing/PathPrinter.hpp"
 #include "mimic++/printing/StatePrinter.hpp"
 #include "mimic++/printing/TypePrinter.hpp"
+#include "mimic++/printing/state/C++23Backports.hpp"
+#include "mimic++/printing/state/CommonTypes.hpp"
 
 #endif
